requirements.txt

streamlit==1.28.1
requests>=2.28.0
APScheduler>=3.9.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
vehicle>=0.10.0    # if you’re using the async RDW wrapper
